backend/requirements.txt
django==5.1.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.4.*
python-dotenv==1.0.*
psycopg2-binary==2.9.*
uvicorn[standard]==0.30.*
faster-whisper==1.0.*
pywebpush==2.0.*
weasyprint==62.*
httpx==0.27.*
whitenoise==6.*
Pillow==10.*
